docker/mediawiki/src/requirements.txt
PyYAML
GitPython